requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
django-filter~=24.2
psycopg2~=2.9.9
drf-spectacular~=0.27.2
djangorestframework-simplejwt~=5.3.1